rowSwapV2

+++ misc/swapType.h
@@ -145,7 +145,7 @@ void swapType(const std::string& source, D& dest) {
 #if QT_VERSION_MAJOR >= 6
 		auto qb = QByteArrayView(source.data(), source.length());
 #else
-		auto qb = QByteArray(source.data(), source.length());
+		auto qb = QByteArray(source.data(), (int)source.length());
 #endif
 		if constexpr (std::is_floating_point_v<D>) {
 			dest = qb.toDouble(&ok);
